genai-usecases/gemini-nano-banana/requirements.txt
google-genai>=1.31.0
python-dotenv>=1.0.1
streamlit>=1.48.1